can use CollectorOutputSlim messages and refactored
